# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClassPulse AI — Real-time classroom intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/{reportId}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a single AI report by ID
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reports/{reportId}/pdf:
    post:
      operationId: renderReportPdf
      tags: [reports]
      summary: Render the report to PDF and store it in Supabase
      description: |
        Drives a headless Chromium against the frontend `/reports/{id}/print`
        route, captures a single-page PDF and uploads it. Returns the public
        URL. Idempotent — re-rendering replaces the prior file at the same path.
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF rendered and stored
          content:
            application/json:
              schema:
                type: object
                required: [pdfUrl]
                properties:
                  pdfUrl:
                    type: string
                    format: uri
        "500":
          description: Render or upload failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reports/{reportId}/send:
    post:
      operationId: sendReportEmail
      tags: [reports]
      summary: Render PDF if needed and email it to the teacher via n8n
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                recipientEmail:
                  type: string
                  format: email
                  description: Override the teacher's on-file email
      responses:
        "200":
          description: Delivery dispatched
          content:
            application/json:
              schema:
                type: object
                required: [status, pdfUrl]
                properties:
                  status:
                    type: string
                    enum: [sent, skipped, failed]
                  recipient:
                    type: string
                    nullable: true
                  pdfUrl:
                    type: string
                    format: uri
                  error:
                    type: string
                    nullable: true
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Class:
      type: object
      properties:
        id:
          type: integer
        teacherId:
          type: integer
        studentName:
          type: string
        subject:
          type: string
        scheduledAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        platform:
          type: string
          enum: [zoom, google_meet, teams]
        meetingUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [upcoming, in_progress, completed, cancelled]
        grade:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [id, teacherId, studentName, subject, scheduledAt, durationMinutes, platform, status]
...
    Report:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        classId:
          type: integer
        teacherId:
          type: integer
        overallScore:
          type: number
        engagementScore:
          type: number
        voiceClarityScore:
          type: number
        interactionScore:
          type: number
        noiseLevel:
          type: number
        speakingConfidence:
          type: number
        deadAirSeconds:
          type: number
        internetStability:
          type: number
        understandingScore:
          type: number
          nullable: true
        satisfactionScore:
          type: number
          nullable: true
        teacherCompatibilityScore:
          type: number
          nullable: true
        churnRiskScore:
          type: number
          nullable: true
        aiSummary:
          type: string
        suggestions:
          type: array
          items:
            type: string
        highlights:
          type: array
          items:
            type: string
        improvementAreas:
          type: array
          items:
            type: string
        timelineData:
          type: array
          items:
            $ref: "#/components/schemas/TimelinePoint"
        moodTimeline:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/MoodPoint"
        confusionTimeline:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/ConfusionPoint"
        createdAt:
          type: string
          format: date-time
        class:
          $ref: "#/components/schemas/Class"
          nullable: true
      required: [id, sessionId, classId, teacherId, overallScore, engagementScore, voiceClarityScore, interactionScore, noiseLevel, speakingConfidence, deadAirSeconds, internetStability, aiSummary, suggestions, highlights, improvementAreas, timelineData, createdAt]
    TimelinePoint:
      type: object
      properties:
        minute:
          type: integer
        engagement:
          type: number
        voiceClarity:
          type: number
        noise:
          type: number
      required: [minute, engagement, voiceClarity, noise]
...
    MoodPoint:
      type: object
      properties:
        minute:
          type: integer
        mood:
          type: string
        valence:
          type: number
        energy:
          type: number
      required: [minute, mood, valence, energy]
    ConfusionPoint:
      type: object
      properties:
        minute:
          type: integer
        confusion:
          type: number
      required: [minute, confusion]
...